Remove W3 Total Cache plugin
After the latest deploy we got inexplicable white screens of death on
all pages. Debugging revealed W3 Total Cache to be the cause of this
although we weren't able to detect why.

Since W3 Total Cache is no longer the recommended plugin for
optimizations anyway and the site performs well without we decided to
remove the plugin.

In production the plugin is already disabled and it is safe to remove it
from the source code now.